+++ CHANGELOG.md
@@ -5,12 +5,13 @@
 ### Patch Changes
 
 - Logger enhancements
+- 70791ff: Improve error handling
 
 ## 0.0.117
 
 ### Patch Changes
 
-- Better type handling for metadata where can be any valid JSON object
+- Better type handling for metadata where can be any valid JSON object ([#110](https://github.com/agentuity/sdk-js/pull/110))
 
 ## 0.0.116